Move slurp_file() into library as jv_load_file()
Needed as part of creating a libjq.
