--- main.h
@@ -1,9 +1,0 @@
-#ifndef MAIN_H
-#define MAIN_H
-
-#include "compile.h"
-
-jv slurp_file(const char*, int);
-
-
-#endif
